Repartitioned code, changed internal structures.
