+++ README.md
@@ -109,6 +109,10 @@ A GUI is also available to let a user fill in configuration options using a form
 
     python3 RedfishServiceValidatorGui.py
 
+To run unittests, use the command:
+    
+    python3 -m unittest tests/*.py
+
 ## Execution flow
 
 1. Redfish Service Validator starts with the Service root Resource Schema by querying the service with the service root URI and getting all the device information, the resources supported and their links. Once the response of the Service root query is verified against its schema, the tool traverses through all the collections and Navigation properties returned by the service.
