+++ cache/README.md
@@ -1,0 +1,11 @@
+To use cache, place modified files into this directory similar to Redfish mockup directory structure.
+
+To replace your ServiceRoot, place 'index.json' to './cache/redfish/v1/', then use command line options:
+
+
+
+`python3 RedfishServiceValidator.py --cache Prefer ./cache`
+
+
+
+Or the config file option listed in config.ini
