+++ tests/simpleeval.py
@@ -10,7 +10,10 @@ from unittest import TestCase
 from unittest import mock
 import datetime
 
-import RedfishServiceValidator as rsv
+import sys
+sys.path.append('../')
+
+from simpletypes import *
 
 class ValidatorTest(TestCase):
 
@@ -26,7 +29,7 @@ class ValidatorTest(TestCase):
         ranges = [empty, empty, (0, None), (None, 21), empty, empty, (0, 15), empty] 
         for num, rang in zip(numbers, ranges):
             minx, maxx = rang
-            paramPass = rsv.validateNumber("TestNum", num,  minx, maxx)
+            paramPass = validateNumber("TestNum", num,  minx, maxx)
             self.assertTrue(paramPass,'This number has failed, {} {}'.format(num,  (minx, maxx)))
 
     def test_validate_int(self):
@@ -35,56 +38,56 @@ class ValidatorTest(TestCase):
         ranges = [empty, empty, (0, None), (None, 21), empty, empty] 
         for num, rang in zip(numbers, ranges):
             minx, maxx = rang
-            paramPass = rsv.validateInt("TestNum", num,  minx, maxx)
+            paramPass = validateInt("TestNum", num,  minx, maxx)
             self.assertTrue(paramPass == isinstance(num, int),'This number has failed, {} {}'.format(num,  (minx, maxx)))
     
     def test_validate_string(self):
         name, string, pattern = "tst", "TestString", None
-        self.assertTrue(rsv.validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
+        self.assertTrue(validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
         name, string, pattern = "tst", "Test_String", "Test.String"
-        self.assertTrue(rsv.validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
+        self.assertTrue(validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
         name, string, pattern = "tst", "100.97.143.52", "^(?:[0-9]{1,3}\.){3}[0-9]{1,3}$"
-        self.assertTrue(rsv.validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
+        self.assertTrue(validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
     
     def test_validate_badstring(self):
         name, string, pattern = "tst", 1, None
-        self.assertFalse(rsv.validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
+        self.assertFalse(validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
         name, string, pattern = "tst", "Test__String", "Test.String"
-        self.assertFalse(rsv.validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
+        self.assertFalse(validateString(name, string, pattern),'This string/pattern has failed, {} {}'.format(string,pattern))
     
     def test_validate_datetime(self):
         name, string = 'date', "2017-05-11T16:30:46-05:00"
-        self.assertTrue(rsv.validateDatetime(name, string),'This datetime has failed: {}'.format(string))
+        self.assertTrue(validateDatetime(name, string),'This datetime has failed: {}'.format(string))
         name, string = 'baddate', "baddate"
-        self.assertFalse(rsv.validateDatetime(name, string),'This datetime has failed: {}'.format(string))
+        self.assertFalse(validateDatetime(name, string),'This datetime has failed: {}'.format(string))
          
     def test_validate_guid(self):
         name, string = 'guid', "deadb33f-0048-3310-8039-01004f003432" 
-        self.assertTrue(rsv.validateGuid(name, string),'This guid has failed: {}'.format(string))
+        self.assertTrue(validateGuid(name, string),'This guid has failed: {}'.format(string))
         name, string = 'badguid', "badguid"
-        self.assertFalse(rsv.validateGuid(name, string),'This guid has failed: {}'.format(string))
+        self.assertFalse(validateGuid(name, string),'This guid has failed: {}'.format(string))
     
     def test_validate_enum(self):
         enums = ['OK','Not OK']
         name, string = 'enum', 'OK'
-        self.assertTrue(rsv.validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertTrue(validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
         name, string = 'enum', 'Not OK'
-        self.assertTrue(rsv.validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertTrue(validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
         name, string = 'badenum', 'DNE'
-        self.assertFalse(rsv.validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertFalse(validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
         name, string = 'badenum', 'Ok'
-        self.assertFalse(rsv.validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertFalse(validateEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
     
     def test_validate_denum(self):
         enums = ['OK','Not OK']
         name, string = 'enum', 'OK'
-        self.assertTrue(rsv.validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertTrue(validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
         name, string = 'enum', [{'Member':'Not OK'}]
-        self.assertTrue(rsv.validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertTrue(validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
         name, string = 'badenum', 'DNE'
-        self.assertFalse(rsv.validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertFalse(validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
         name, string = 'badenum', [{'Member':'Ok'}]
-        self.assertFalse(rsv.validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
+        self.assertFalse(validateDeprecatedEnum(name, string, enums),'This enum has failed: {} {}'.format(string, enums))
     
     def test_validate_complex(self):
         self.assertTrue(True,'Huh?')
